Allow swagger-servant-ui with redoc to be nested in an API
In the current state, if we were to declare a servant-swagger-ui server behind some path, for instance:

    type API =
        "docs" :> ServantSwaggerUI "swagger-ui" "swagger.json"

One would correctly have access to the template in `<host>/docs/swagger.ui`
but the template would refer to a swagger.json at `/swagger.json`
whereas it is hosted under `/docs/swagger.json`. To cope with this, we
can use relative paths in the template, making sure that the current url
is always ending with a slash.